Add tini init
Currently there is a problem in which arcanist is taking a long time,
reportedly blocked waiting for some arbitrary event.

This commit introduces tini, making the leap that it's possible arcanist
is looking for a socket for a process that is dead, but not cleaned up
by arcanist itself.

It's a large conceptual leap, but tini should be present as init
regardless.